requirements.txt
chainlit
langchain
langchain-community
langchain-chroma
langchain-ollama
sentence_transformers
Pillow